requirements.txt
selenium == 4.7.2
pytest == 7.2.0
allure-pytest==2.12.0
pytest-xdist==3.1.0
pytest-check==2.1.4
